Update to v103r31 release.
byuu says:

Changelog:

  - gba/cpu: slight speedup to CPU::step()
  - processor/arm7tdmi: fixed about ten bugs, ST018 and GBA games are
    now playable once again
  - processor/arm: removed core from codebase
  - processor/v30mz: code cleanup (renamed functions; updated
    instruction() for consistency with other cores)

It turns out on my much faster system, the new ARM7TDMI core is very
slightly slower than the old one (by about 2% or so FPS.) But the
CPU::step() improvement basically made it a wash.

So yeah, I'm in really serious trouble with how slow my GBA core is now.
Sigh.

As for higan/processor ... this concludes the first phase of major
cleanups and rewrites.

There will always be work to do, and I have two more phases in mind.

One is that a lot of the instruction disassemblers are very old. One
even uses sprintf still. I'd like to modernize them all. Also, the
ARM7TDMI core (and the ARM core before it) can't really disassemble
because the PC address used for instruction execution is not known prior
to calling instruction(), due to pipeline reload fetches that may occur
inside of said function. I had a nasty hack for debugging the new core,
but I'd like to come up with a clean way to allow tracing the new
ARM7TDMI core.

Another is that I'd still like to rename a lot of instruction function
names in various cores to be more descriptive. I really liked how the
LR35902 core came out there, and would like to get that level of detail
in with the other cores as well.

+++ higan/processor/arm7tdmi/memory.cpp
@@ -12,11 +12,11 @@ auto ARM7TDMI::load(uint mode, uint32 address) -> uint32 {
   auto word = get(Load | mode, address);
   if(mode & Half) {
     address &= 1;
-    word = mode & Signed ? (uint16)(int16)word : (uint16)word;
+    word = mode & Signed ? (uint32)(int16)word : (uint32)(uint16)word;
   }
   if(mode & Byte) {
     address &= 0;
-    word = mode & Signed ? (uint8)(int8)word : (uint8)word;
+    word = mode & Signed ? (uint32)(int8)word : (uint32)(uint8)word;
   }
   if(mode & Signed) {
     word = ASR(word, address.bits(0,1) << 3);
